chore(fe): move SideBar to root component
